Update test suite
