+++ sc/tests/bundle.py
@@ -3,13 +3,9 @@ import os
 import sys
 from email.mime.text import MIMEText
 import time
-path = os.path.dirname(os.path.abspath(__file__)).split(os.sep)
-del path[-1]
-path = os.path.normpath(os.sep.join(path))
-if path not in sys.path:
-    sys.path += [path]
-from app.functions import *
-from app.task import Task
+from sc.app.functions import *
+from sc.app.task import Task
+from util import create_fac, create_group, create_user
 
 class SystemTest(unittest.TestCase):
     '''
@@ -887,7 +883,7 @@ class TestFull(unittest.TestCase):
 
         # grab some prepackaged geoJSON to ensure we have some 
         # events and shakemaps for testing
-        json_file = os.path.join(sc_dir(), 'test', 'data', 'test_json_feed.json')
+        json_file = os.path.join(sc_dir(), 'tests', 'data', 'test_json_feed.json')
         with open(json_file, 'r') as file_:
             json_feed = json.loads(file_.read())
         pg = ProductGrabber()
@@ -997,7 +993,7 @@ class TestFull(unittest.TestCase):
 
         # grab some prepackaged geoJSON to ensure we have some 
         # events and shakemaps for testing
-        json_file = os.path.join(sc_dir(), 'test', 'data', 'test_json_feed.json')
+        json_file = os.path.join(sc_dir(), 'tests', 'data', 'test_json_feed.json')
         with open(json_file, 'r') as file_:
             json_feed = json.loads(file_.read())
         pg = ProductGrabber()
@@ -1079,7 +1075,7 @@ class TestFull(unittest.TestCase):
         s = SoftwareUpdater()
         s.json_url = os.path.normpath(delim.join(['file://', 
                                                     sc_dir(), 
-                                                    'test',
+                                                    'tests',
                                                     'data',
                                                     'update_test.json']))
         s.check_update(testing=True)
@@ -1088,7 +1084,7 @@ class TestFull(unittest.TestCase):
         s = SoftwareUpdater()
         s.json_url = os.path.normpath(delim.join(['file://', 
                                                     sc_dir(), 
-                                                    'test',
+                                                    'tests',
                                                     'data',
                                                     'update_test.json']))
         s.update(testing=True)
@@ -1236,7 +1232,7 @@ class TestImport(unittest.TestCase):
 
     @dbconnect
     def step02_userImport(self, session=None):
-        user_file = os.path.join(sc_dir(), 'test', 'data', 'test_users.xml')
+        user_file = os.path.join(sc_dir(), 'tests', 'data', 'test_users.xml')
         file_type = determine_xml(user_file)
         import_user_xml(user_file)
 
@@ -1253,13 +1249,13 @@ class TestImport(unittest.TestCase):
         self.assertEqual(user.mms, 'example@example.com')
 
     def step03_groupImport(self):
-        group_file = os.path.join(sc_dir(), 'test', 'data', 'test_groups.xml')
+        group_file = os.path.join(sc_dir(), 'tests', 'data', 'test_groups.xml')
         file_type = determine_xml(group_file)
         import_group_xml(group_file, 1)
         self.assertEqual(file_type, 'group')
         
     def step04_facImport(self):
-        fac_file = os.path.join(sc_dir(), 'test', 'data', 'test_facs.xml')
+        fac_file = os.path.join(sc_dir(), 'tests', 'data', 'test_facs.xml')
         file_type = determine_xml(fac_file)
         import_facility_xml(fac_file, 1)
         self.assertEqual(file_type, 'facility')
@@ -1377,7 +1373,7 @@ class TestImport(unittest.TestCase):
         self.step01_clearData()
 
     def step12_masterImport(self):
-        file_ = os.path.join(sc_dir(), 'test', 'data', 'test_master.xml')
+        file_ = os.path.join(sc_dir(), 'tests', 'data', 'test_master.xml')
         file_type = determine_xml(file_)
         import_master_xml(file_)
 
@@ -1524,91 +1520,6 @@ class TestSCConfig(unittest.TestCase):
             # fail the test with the accumulated exception message
             self.fail(fail_message)
 
-def create_fac(grid=None, fac_id='AUTO_GENERATED'):
-    '''
-    Create a facility that is inside of a grid with generic fragility
-    '''
-    
-    facility = Facility()
-    if grid:
-        lat_adjust = abs((grid.lat_max - grid.lat_min) / 10)
-        lon_adjust = abs((grid.lon_max - grid.lon_min) / 10)
-        facility.lat_min = grid.lat_min + lat_adjust
-        facility.lat_max = facility.lat_min + (2 * lat_adjust)
-        facility.lon_min = grid.lon_min + lon_adjust
-        facility.lon_max = facility.lon_min + (2 * lon_adjust)
-    
-    facility.facility_id = fac_id
-    facility.facility_type = 'Bridge'
-    facility.name = 'No Name'
-    facility.metric = 'MMI'
-    facility.grey = 0
-    facility.green = 3
-    facility.yellow = 5
-    facility.orange = 6
-    facility.red = 7
-    facility.grey_beta = .64
-    facility.green_beta = .64
-    facility.yellow_beta = .64
-    facility.orange_beta = .64
-    facility.red_beta = .64
-    
-    return facility
-    
-def create_group(name=None, 
-                    event_type='ACTUAL',
-                    notification_format=None,
-                    new_event=True,
-                    heartbeat=True,
-                    insp_prios=['GREY', 
-                                'GREEN', 
-                                'YELLOW', 
-                                'ORANGE', 
-                                'RED']):
-    group = Group()
-    group.name = name
-    group.facility_type = 'All'
-    group.lon_min = -180
-    group.lon_max = 180
-    group.lat_min = -90
-    group.lat_max = 90
-    
-    if new_event is True:
-        gs = GroupSpecification()
-        gs.event_type = event_type
-        gs.notification_type = 'NEW_EVENT'
-        gs.minimum_magnitude = 3
-        gs.notification_format = notification_format
-        group.specs.append(gs)
-    
-    if heartbeat is True:
-        gs = GroupSpecification()
-        gs.event_type = event_type
-        gs.notification_type = 'new_event'
-        gs.notification_format = notification_format
-        gs.event_type = 'heartbeat'
-        group.specs.append(gs)
-    
-    for insp_prio in insp_prios:
-        gs = GroupSpecification()
-        gs.event_type = event_type
-        gs.notification_type = 'DAMAGE'
-        gs.minimum_magnitude = 3
-        gs.notification_format = notification_format
-        gs.inspection_priority = insp_prio
-        group.specs.append(gs)
-
-    return group
-
-def create_user(group_str=None, email=None, mms=None):
-    user = User()
-    user.username = 'test_user'
-    user.email = email
-    user.mms = mms
-    user.user_type = 'ADMIN'
-    user.group_string = group_str
-
-    return user
 
 def fail_test():
     '''
